backend/requirements.txt
Flask
python-dotenv
cryptography
flask-cors
africastalking
pyrebase4
pycryptodome